코드 재검수 전 수정사항
*** 클라이언트
* x64 환경으로 수정
* void채널로 이동된 경우 edit 박스 표시 ("Channel Disconnected")

*** 더미 클라이언트
* x64 환경으로 수정
* Send 및 랜덤패킷 상황에서의 new OverlapExp에서 정적으로 변경
	미리 std::vector 생성, concurrent_queue에서 인덱스 관리

*** 서버
* x64 환경으로 수정
* Send 상황에서의 new OverlapExp를 더미 클라이언트와 같이 변경
* 클라이언트 락 작성, 기존에 클라이언트 락 없이 동기화가 가능했던
  로그인, 종료, 채널리스트를 제외하고
  채널변경 및 강퇴 상황에서 필요한 곳에 부분적으로 사용

코드 검수, 주석 작성 이후 재 테스트 진행, 마무리 후 원격 테스트

+++ ChattingServer/Channel.h
@@ -16,7 +16,7 @@ public:
 	virtual void	Exit(const Client* client) = 0;
 	
 
-	unsigned int						GetUserCount() const { return clientsInChannel.size(); }
+	size_t								GetUserCount() const { return clientsInChannel.size(); }
 	std::string							GetChannelName() const { return channelName; }
 	std::string							GetChannelMaster() const { return channelMaster; }
 	std::mutex&							GetChannelLock() { return channelLock; }
